revise fish case
